Renamed ZoomMotion to ZoomShared
Confusing having motion in the name
