+++ NavigationReactNative/sample/zoom/ZoomShared.js
